ACPI / LPSS: Only add device links from consumer to supplier
Commit e6ce0ce34f65 ("ACPI / LPSS: Add device link for CHT SD card
dependency on I2C") tries to add device links from consumer to
supplier *and* the other way around.

When the consumer gets registered the supplier is not yet registered, so
acpi_lpss_find_device() cannot find it and we never end up adding the
supplier link. Although not intended this is a good thing, because the
device links are dependencies and if we add a link in both directions the
power-management core will not know which one to suspend/resume first.

I've verified through debug printk-s that indeed only the consumer
to supplier dependency gets added even before this commit removes
the code to add the link the other way.

Tested-by: Jarkko Nikula <jarkko.nikula@linux.intel.com>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>