platform/x86: intel_int0002_vgpio: Enable the driver on Bay Trail platforms
The intel_int0002_vgpio driver was added to avoid an IRQ 9 storm on Cherry
Trail platforms.

When originally merged the CPU ID for Bay Trail SoCs was commented out
of the list of valid CPU IDs because we did not have any reports of the
IRQ storm on Bay Trail platforms.

We now have a report of the IRQ 9 storm on the Bay Trail based Thinkpad
Tablet 10 which is fixed by enabling this driver, so lets enable it on
Bay Trail too.

I've tested various other Bay Trail device with this driver enabled
without adverse side-effects.

BugLink: https://www.dpin.de/nf/finally-s0i3-is-there-thinkpad-tablet-10-sleeps-deeply-with-linux-kernel-4-15rc/#comments
Reported-and-tested-by: Nicole Faerber <nicole@id3p.de>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>