pwm: lpss: Restore last settings on resume, rather then save + restore regs
On some Cherry Trail devices the AML _PS0 and _PS3 methods of the GPU
(GFX0 device) are directly poking the PWM controller leaving it in a
different mode then we configured it.

Since the PWM is already enabled by the _PS0 method of the GPU (which runs
in the noirq resume phase), restoring the old register settings is not
enough, we must set the SW_UPDATE flag to get the bad settings done by the
GPU's _PS0 method replaced with the new settings.

This commit replaces the suspend/resume doing a save/restore with a new
resume handler which reprograms the pwm controller with the last settings
as stored in pwm_device.state, properly setting the update bit to apply
the new settings.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>